daemon-cli: move connect verb into daemon-cli crate
Moves connect from the binary into doublezero-daemon-cli per RFC-20
(PR 6 of the extraction stack, malbeclabs/infra#1492). Device selection
now uses the crate's latency utilities; the binary's dzd_latency.rs
(connect was its last user) and the orphaned check_doublezero pre-flight
are deleted, and servicecontroller.rs is trimmed to the v2_status
surface the multicast verbs still need.
