+++ CHANGELOG.md
@@ -38,6 +38,7 @@ All notable changes to this project will be documented in this file.
 - Onchain programs
   - Restrict granting the `FOUNDATION` permission flag: a plain `PERMISSION_ADMIN` holder can no longer grant `FOUNDATION` (a privilege escalation). Only a `foundation_allowlist` member or an existing `FOUNDATION` holder may grant it, enforced independently of `RequirePermissionAccounts` so foundation members are never locked out.
 - CLI
+  - Move the `connect` verb into the `doublezero-daemon-cli` crate per RFC-20. It now takes `&CliContext` + generic `&D: DaemonClient` + `&L: LedgerClient` + `&mut W` writer; informational/result lines route through the shared writer, spinners stay on stderr, and device selection uses the crate's latency utilities (from #3995). The binary's `dzd_latency.rs` and the orphaned `check_doublezero` pre-flight are deleted. Behavior (flags, output, `--verbose`, version-check) is unchanged.
   - Move the `disconnect` verb into the `doublezero-daemon-cli` crate per RFC-20. It now takes `&CliContext` + generic `&D: DaemonClient` + `&L: LedgerClient` + `&mut W` writer; informational/result lines route through the shared writer, spinners stay on stderr, and diagnostics route through `tracing`. Behavior (flags, output, `--verbose`, `--no-wait`, version-check) is unchanged. (#4008)
   - Move the `latency` and `routes` verbs, plus the device-selection/latency-polling utilities and `resolve_client_ip`, into the `doublezero-daemon-cli` crate per RFC-20. Both verbs now take `&CliContext` + generic `&D: DaemonClient` + `&L: LedgerClient` + `&mut W` writer; diagnostics route through `tracing` and output through the shared writer helper. Behavior (flags, output, `--json`, version-check) is unchanged. (#3995)
   - Add `doublezero permission audit` to check legacy→Permission parity before enabling `require-permission-accounts`: it reports which legacy keys would lose access to migrated instructions (coverage gaps, non-zero exit), super-admin holders, and the non-migrated subsystems that still depend on the GlobalState allowlists.
