+++ client/doublezero/src/cli/command.rs
@@ -4,27 +4,21 @@ use doublezero_daemon_cli::DaemonCommand;
 use doublezero_geolocation_cli::GeolocationArgs;
 use doublezero_serviceability_cli::cli::ServiceabilityCommand;
 
-use crate::{
-    cli::{multicast::MulticastCliCommand, sentinel::SentinelCliCommand},
-    command::connect::ProvisioningCliCommand,
-};
+use crate::cli::{multicast::MulticastCliCommand, sentinel::SentinelCliCommand};
 
 /// Top-level command tree for the unified `doublezero` binary.
 ///
 /// Per RFC-20 §Module contract item 2, module-crate verbs are hoisted to the
 /// top level via `#[command(flatten)]`: serviceability verbs from
-/// `doublezero_serviceability_cli`, daemon-control verbs (`enable`, `disable`,
-/// `status`, `disconnect`, `latency`, `routes`) from `doublezero_daemon_cli`.
-/// The binary retains the not-yet-migrated
-/// daemon-control verbs, the `doublezero-geolocation-cli` module crate's
-/// geolocation subtree (via `GeolocationArgs`), the binary-only `Completion`
-/// generator, and `Multicast` (whose `Subscribe`/`Unsubscribe`/`Publish`/
-/// `Unpublish` arms depend on binary-local daemon-control infrastructure).
+/// `doublezero_serviceability_cli`, daemon-control verbs (`connect`, `enable`,
+/// `disable`, `status`, `disconnect`, `latency`, `routes`) from
+/// `doublezero_daemon_cli`. The binary retains the
+/// `doublezero-geolocation-cli` module crate's geolocation subtree (via
+/// `GeolocationArgs`), the binary-only `Completion` generator, and `Multicast`
+/// (whose `Subscribe`/`Unsubscribe`/`Publish`/`Unpublish` arms depend on
+/// binary-local daemon-control infrastructure).
 #[derive(Subcommand, Debug)]
 pub enum Command {
-    /// Connect your server to a doublezero device
-    Connect(ProvisioningCliCommand),
-
     /// Daemon-control verbs migrated to `doublezero-daemon-cli` (RFC-20).
     /// Hoisted to top-level via `#[command(flatten)]`.
     #[command(flatten)]
