+++ client/doublezero/src/command/mod.rs
@@ -1,3 +1,2 @@
-pub mod connect;
 pub mod helpers;
 pub mod multicast;
